// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.4"
  }
  public: {
    Tables: {
      accounts: {
        Row: {
          balance: number
          created_at: string
          id: string
          name: string
          type: string
          user_id: string | null
        }
        Insert: {
          balance?: number
          created_at?: string
          id?: string
          name: string
          type: string
          user_id?: string | null
        }
        Update: {
          balance?: number
          created_at?: string
          id?: string
          name?: string
          type?: string
          user_id?: string | null
        }
        Relationships: []
      }
      categories: {
        Row: {
          created_at: string
          id: string
          name: string
          type: string
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: string
          name: string
          type: string
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: string
          name?: string
          type?: string
          user_id?: string | null
        }
        Relationships: []
      }
      credit_cards: {
        Row: {
          closing_day: number
          created_at: string
          due_day: number
          id: string
          limit: number
          name: string
          user_id: string | null
        }
        Insert: {
          closing_day: number
          created_at?: string
          due_day: number
          id?: string
          limit: number
          name: string
          user_id?: string | null
        }
        Update: {
          closing_day?: number
          created_at?: string
          due_day?: number
          id?: string
          limit?: number
          name?: string
          user_id?: string | null
        }
        Relationships: []
      }
      goals: {
        Row: {
          created_at: string
          current_amount: number
          deadline: string
          id: string
          name: string
          target_amount: number
          user_id: string | null
        }
        Insert: {
          created_at?: string
          current_amount?: number
          deadline: string
          id?: string
          name: string
          target_amount: number
          user_id?: string | null
        }
        Update: {
          created_at?: string
          current_amount?: number
          deadline?: string
          id?: string
          name?: string
          target_amount?: number
          user_id?: string | null
        }
        Relationships: []
      }
      investments: {
        Row: {
          amount: number
          created_at: string
          date: string
          id: string
          name: string
          type: string
          user_id: string | null
          yield: number
        }
        Insert: {
          amount: number
          created_at?: string
          date: string
          id?: string
          name: string
          type: string
          user_id?: string | null
          yield?: number
        }
        Update: {
          amount?: number
          created_at?: string
          date?: string
          id?: string
          name?: string
          type?: string
          user_id?: string | null
          yield?: number
        }
        Relationships: []
      }
      transactions: {
        Row: {
          account_id: string | null
          amount: number
          category_id: string | null
          created_at: string
          credit_card_id: string | null
          date: string
          description: string
          fixed_end_date: string | null
          id: string
          installments: number | null
          is_fixed: boolean | null
          is_paid: boolean | null
          liters: number | null
          mileage: number | null
          payment_method: string | null
          type: string
          user_id: string | null
          vehicle_id: string | null
        }
        Insert: {
          account_id?: string | null
          amount: number
          category_id?: string | null
          created_at?: string
          credit_card_id?: string | null
          date: string
          description: string
          fixed_end_date?: string | null
          id?: string
          installments?: number | null
          is_fixed?: boolean | null
          is_paid?: boolean | null
          liters?: number | null
          mileage?: number | null
          payment_method?: string | null
          type: string
          user_id?: string | null
          vehicle_id?: string | null
        }
        Update: {
          account_id?: string | null
          amount?: number
          category_id?: string | null
          created_at?: string
          credit_card_id?: string | null
          date?: string
          description?: string
          fixed_end_date?: string | null
          id?: string
          installments?: number | null
          is_fixed?: boolean | null
          is_paid?: boolean | null
          liters?: number | null
          mileage?: number | null
          payment_method?: string | null
          type?: string
          user_id?: string | null
          vehicle_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "transactions_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "transactions_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "transactions_credit_card_id_fkey"
            columns: ["credit_card_id"]
            isOneToOne: false
            referencedRelation: "credit_cards"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "transactions_vehicle_id_fkey"
            columns: ["vehicle_id"]
            isOneToOne: false
            referencedRelation: "vehicles"
            referencedColumns: ["id"]
          },
        ]
      }
      vehicles: {
        Row: {
          created_at: string
          id: string
          name: string
          plate: string
          user_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          name: string
          plate: string
          user_id: string
        }
        Update: {
          created_at?: string
          id?: string
          name?: string
          plate?: string
          user_id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}